Rename PriceCalculator -> PriceCalculatorForm
