+++ apps/store/src/components/PriceCalculatorForm/StepIconValid.tsx
@@ -1,6 +1,6 @@
 import { useTheme } from '@emotion/react'
 import styled from '@emotion/styled'
-import { STEP_ICON_SIZE } from './PriceCalculator.constants'
+import { STEP_ICON_SIZE } from './PriceCalculatorForm.constants'
 import { TickIcon } from './TickIcon'
 
 const StyledStep = styled.div(({ theme }) => ({
