+++ apps/store/src/components/PriceCalculatorForm/TickIcon.tsx
